feature: customer, voucher entity 구현
