--- src/main/java/com/programmers/springbasic/entity/Customer.java
@@ -1,16 +1,0 @@
-package com.programmers.springbasic.entity;
-
-public class Customer {
-	private Long id;
-	private String name;
-	private Boolean isBlacklisted;
-
-	public Long getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-}
