+++ src/main/java/com/programmers/springbasic/entity/customer/Customer.java
@@ -1,0 +1,25 @@
+package com.programmers.springbasic.entity.customer;
+
+public class Customer {
+	private Long id;
+	private String name;
+	private Boolean isBlacklisted;
+
+	public Customer(Long id, String name, Boolean isBlacklisted) {
+		this.id = id;
+		this.name = name;
+		this.isBlacklisted = isBlacklisted;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public Boolean getBlacklisted() {
+		return isBlacklisted;
+	}
+}
